+++ include/boid.hpp
@@ -8,30 +8,30 @@ namespace boids_sim
 
     struct Vector2D
     {
-        double x;
-        double y;
+        float x;
+        float y;
 
         Vector2D operator+(const Vector2D &other) const;
         Vector2D operator-(const Vector2D &other) const;
-        Vector2D operator*(double scalar) const;
+        Vector2D operator*(float scalar) const;
         Vector2D &operator+=(const Vector2D &other);
-        double norm2() const;
+        float norm2() const;
     };
     class boid
     {
     private:
-        Vector2D position_{0.0, 0.0};
-        Vector2D velocity_{0.0, 0.0};
+        Vector2D position_{0.f, 0.f};
+        Vector2D velocity_{0.f, 0.f};
         
 
     public:
         boid() = default;
-        boid(double x, double y, double vx, double vy);
+        boid(float x, float y, float vx, float vy);
 
         const Vector2D &getPosition() const;
         const Vector2D &getVelocity() const;
 
-        void updatepositon(double dt, double maxX, double maxY);
+        void updatepositon(float dt, float maxX, float maxY);
         void updatevelocity();
     };
 }
